ref: split executor into its own module
